\name{SensitivityFunctions-package}
\alias{SensitivityFunctions-package}
\alias{SensitivityFunctions}
\docType{package}
\title{
\packageTitle{SensitivityFunctions}
}
\description{
\packageDescription{SensitivityFunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SensitivityFunctions}
\packageIndices{SensitivityFunctions}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SensitivityFunctions}

Maintainer: \packageMaintainer{SensitivityFunctions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
